Now compiles with Visual C++ :-D
IOLib: types corrected.
MathLib: "#define _USE_MATH_DEFINES" added so that M_PI is defined.
StdLib: getOS() added to remove preprocessor usage in function macro.
Win32Lib: ...
EScript.h: annoying warning disabled.
Bool: "+" and "-" removed (hope these were not used anywhere)
Map: warning removed.
Number: 
 - "#define _USE_MATH_DEFINES" added so that M_PI is defined.
 - types adapted
Object: warning removed.
Parser:
 - annoying warning disabled.
 - types adapted.
Token: warning removed.
FileUtils: 
 - types adapted.
 - msvc-compatible dirent.h wrapper added.
Macros: 
 - only partly supported __attribute__ made optional
StringUtils: warning removed.


 

git-svn-id: https://svn.berlios.de/svnroot/repos/escript/branches/0.5.4.Stoepsel@51 5693b1fc-3b75-4070-9b6b-3ce3692a40d5

+++ EScript/Utils/FileUtils.h
@@ -10,8 +10,8 @@
 namespace EScript {
 namespace FileUtils{
 
-char * loadFile(const std::string & filename,long & size);
-bool saveFile(const std::string & filename,const char * content,const long size);
+char * loadFile(const std::string & filename,size_t & size);
+bool saveFile(const std::string & filename,const char * content,const size_t size);
 
 /*! @param filename
  *	@return file modification Time	*/
